Move ads classes to their correct packages
