+++ build.gradle
@@ -3,10 +3,11 @@
 buildscript {
     repositories {
         jcenter()
+        google()
     }
     dependencies {
-        classpath 'com.android.tools.build:gradle:2.3.2'
-        classpath 'com.novoda:bintray-release:0.3.4'
+        classpath 'com.android.tools.build:gradle:3.0.0'
+        classpath 'com.novoda:bintray-release:0.5.0'
     }
 }
 
@@ -17,6 +18,7 @@ allprojects {
         /*maven {
             url 'https://jitpack.io'
         }*/
+        google()
     }
 }
