+++ playkit/build.gradle
@@ -3,7 +3,7 @@ apply from: 'version.gradle'
 
 android {
     compileSdkVersion 25
-    buildToolsVersion "25.0.2"
+    buildToolsVersion '26.0.2'
 
     defaultConfig {
         minSdkVersion 16
@@ -32,7 +32,6 @@ tasks.withType(Javadoc) {
 }
 
 dependencies {
-    compile project(":admanager")
     compile 'com.kaltura.netkit:netkit-core:1.0.5'
     //compile ('com.kaltura.netkit-android:netkit-core:b6175662cfae258e98a5b028363fbf0962d9c5a8')
 
@@ -40,12 +39,12 @@ dependencies {
     compile 'com.google.android.exoplayer:exoplayer:r2.3.1'
 
     // Google Ads
-    compile 'com.google.ads.interactivemedia.v3:interactivemedia:3.7.2'
-    compile 'com.google.android.gms:play-services-ads:11.0.4'
+    compile 'com.google.ads.interactivemedia.v3:interactivemedia:3.7.4'
+    compile 'com.google.android.gms:play-services-ads:11.4.2'
 
     // Google Cast
-    compile 'com.google.android.gms:play-services-cast-framework:11.0.4'
-    compile 'com.android.support:mediarouter-v7:25.3.1'
+    compile 'com.google.android.gms:play-services-cast-framework:11.4.2'
+    compile 'com.android.support:mediarouter-v7:25.4.0'
 
     // Youbora Analytics
     compile files('lib/YouboraLib-5.3.10.jar')
@@ -54,6 +53,6 @@ dependencies {
     testCompile 'junit:junit:4.12'
     testCompile 'org.hamcrest:hamcrest-library:1.3'
     testCompile "org.mockito:mockito-core:2.2.26"
-    androidTestCompile 'com.android.support.test:runner:0.5'
+    androidTestCompile 'com.android.support.test:runner:1.0.1'
 }
