+++ playkit/src/main/java/com/kaltura/playkit/ads/AdEnabledPlayerController.java
@@ -17,7 +17,6 @@ import android.support.annotation.NonNull;
 import com.kaltura.playkit.PKLog;
 import com.kaltura.playkit.PKMediaConfig;
 import com.kaltura.playkit.PlayerDecorator;
-import com.kaltura.playkit.plugins.ads.AdsProvider;
 import com.kaltura.playkit.utils.Consts;
 
 /**
