+++ playkit/src/main/java/com/kaltura/playkit/plugins/ovp/KalturaStatsPlugin.java
@@ -30,8 +30,8 @@ import com.kaltura.playkit.PlayKitManager;
 import com.kaltura.playkit.Player;
 import com.kaltura.playkit.PlayerEvent;
 import com.kaltura.playkit.api.ovp.services.StatsService;
-import com.kaltura.playkit.plugins.ads.AdEvent;
-import com.kaltura.playkit.plugins.ads.AdPositionType;
+import com.kaltura.playkit.ads.AdEvent;
+import com.kaltura.playkit.ads.AdPositionType;
 import com.kaltura.playkit.utils.Consts;
 
 import java.util.TimerTask;
