+++ playkit/src/main/java/com/kaltura/playkit/plugins/youbora/YouboraAdManager.java
@@ -20,8 +20,8 @@ import com.kaltura.playkit.PlayKitManager;
 import com.kaltura.playkit.PlayerEvent;
 import com.kaltura.playkit.ads.PKAdEndedReason;
 import com.kaltura.playkit.ads.PKAdErrorType;
-import com.kaltura.playkit.plugins.ads.AdEvent;
-import com.kaltura.playkit.plugins.ads.AdPositionType;
+import com.kaltura.playkit.ads.AdEvent;
+import com.kaltura.playkit.ads.AdPositionType;
 import com.kaltura.playkit.utils.Consts;
 import com.npaw.youbora.adnalyzers.AdnalyzerGeneric;
 import com.npaw.youbora.plugins.PluginGeneric;
@@ -29,7 +29,7 @@ import com.npaw.youbora.youboralib.BuildConfig;
 import com.npaw.youbora.youboralib.utils.YBLog;
 
 import static com.kaltura.playkit.PlayerEvent.Type.STATE_CHANGED;
-import static com.kaltura.playkit.plugins.ads.AdPositionType.UNKNOWN;
+import static com.kaltura.playkit.ads.AdPositionType.UNKNOWN;
 
 /**
  * @hide
