+++ playkitdemo/src/main/java/com/kaltura/playkitdemo/MainActivity.java
@@ -50,8 +50,8 @@ import com.kaltura.playkit.player.PKTracks;
 import com.kaltura.playkit.player.TextTrack;
 import com.kaltura.playkit.player.VideoTrack;
 import com.kaltura.playkit.plugins.SamplePlugin;
-import com.kaltura.playkit.plugins.ads.AdCuePoints;
-import com.kaltura.playkit.plugins.ads.AdEvent;
+import com.kaltura.playkit.ads.AdCuePoints;
+import com.kaltura.playkit.ads.AdEvent;
 import com.kaltura.playkit.plugins.ads.ima.IMAConfig;
 import com.kaltura.playkit.plugins.ads.ima.IMAPlugin;
 import com.kaltura.playkit.plugins.ads.kaltura.ADConfig;
