+++ settings.gradle
@@ -1,2 +1,1 @@
-include ':playkit', ':admanager', ':playkitdemo'
-project(':admanager').projectDir = new File('../../playkit-admanager-android/admanager')
+include ':playkit', ':playkitdemo'
